cluster: extract session-gen guards into sync_conn_gen.go
Pure code-motion split of pkg/cluster/sync_conn.go (2228 lines, over the
~2000 modularity threshold tracked in #5661). This first increment moves
the HA session generation-guard and synced-session apply machinery out to
a cohesive sibling file in the same package.

Moved verbatim (no rename, no logic/timing change):
  noteHelperMirrorResult, genGuardMapCap, putGenBounded, nextInstallGen,
  fullSetSeqGuard (+admit/reset), stampInstallGenV4/V6, takeDeleteGenV4/V6,
  installGenGuardV4/V6, recordInstalledGenV4/V6, deleteGenGuardV4/V6,
  resetRecvGen, configEpochStale, installClusterSyncedV4/V6,
  deleteClusterSyncedV4/V6.

Same package (pkg/cluster) so no symbol renames or visibility changes are
needed; imports recomputed per file. Behavior is byte-identical.

go build ./... passes.
