+++ pkg/cluster/sync_conn.go
@@ -14,494 +14,6 @@ import (
 	"github.com/psaab/xpf/pkg/dataplane"
 )
 
-func (s *SessionSync) noteHelperMirrorResult(af string, warned *atomic.Bool, err error) {
-	if err == nil {
-		warned.Store(false)
-		return
-	}
-	s.stats.Errors.Add(1)
-	if warned.CompareAndSwap(false, true) {
-		slog.Warn("cluster sync: failed to mirror synced session into dataplane helper", "af", af, "err", err)
-		return
-	}
-	slog.Debug("cluster sync: repeated synced-session helper mirror failure", "af", af, "err", err)
-}
-
-// genGuardMapCap bounds the sender-side echo maps and the receiver-side
-// stored-generation maps so a churning workload cannot grow them without
-// limit. Both are evicted on delete; the cap is a safety valve for keys whose
-// delete never arrives (e.g. dropped close delta). It matches the delete
-// journal cap order-of-magnitude.
-//
-// Overflow handling (#2198 F1): when a map is at cap, a NEW key is NOT
-// recorded (skip-record-on-full) and an EXISTING key is updated in place. The
-// map is NEVER cleared. Clearing the whole map would drop the stored
-// generation of every live key, disabling the guard cluster-wide for a churn
-// window — exactly the #2170 hazard the guard exists to close: a stale delete
-// could then kill a live re-established session. A skipped new key degrades to
-// gen-0 (unconditional delete / unconditional install), which is always SAFE
-// — gen-0 never causes a wrongful delete of a *different* live incarnation,
-// it only loses the stale-delete protection for that one new key.
-const genGuardMapCap = 200000
-
-// putGenBounded records gen for key in m without ever clearing the map or
-// dropping the stored generation of an existing key. An existing key is always
-// updated in place; a new key is recorded only while the map is below
-// genGuardMapCap. Returns true if the entry was stored. The caller holds the
-// map's mutex. Generic over the two wire-key types.
-func putGenBounded[K comparable](m map[K]uint64, key K, gen uint64) bool {
-	if _, exists := m[key]; exists {
-		m[key] = gen
-		return true
-	}
-	if len(m) >= genGuardMapCap {
-		// Map is full and this is a new key: skip-record. The key degrades to
-		// gen-0 behavior, which is safe (see genGuardMapCap doc).
-		return false
-	}
-	m[key] = gen
-	return true
-}
-
-// nextInstallGen returns the next strictly-monotonic install generation.
-func (s *SessionSync) nextInstallGen() uint64 {
-	return s.genCounter.Add(1)
-}
-
-// fullSetSeqGuard is the receiver-side high-water mark for one FULL-SET sync
-// stream (IPsec SA, or DHCP leases per family) under #5706. It records the
-// last-applied (incarnation, seq) and admits only a strictly-newer pair, so a
-// stale full-set delivered out of order across the redundant fabric
-// receiveLoops (conn0/conn1) is dropped instead of regressing the held set.
-//
-// Ordering is lexicographic on (incarnation, seq): a strictly-higher
-// incarnation always supersedes (a peer restart draws a fresh epoch), and
-// within an incarnation the per-type seq orders the pushes. A LOWER
-// incarnation is stale and refused — EXCEPT that the guard is reset() to zero
-// on a peer bulk re-prime (resetRecvGen), so an OS-rebooted peer whose
-// monotonic epoch restarts LOWER is re-accepted from its fresh set rather than
-// stranded on the pre-reboot set (the #2198 F2 stale-RETAIN inverse, applied
-// to full-set sync). The zero value is the initial/legacy state.
-//
-// Access is serialized by SessionSync.recvSeqMu because the two fabric
-// receiveLoops can invoke admit concurrently.
-type fullSetSeqGuard struct {
-	incarnation uint64
-	seq         uint64
-}
-
-// admit reports whether a full-set stamped (incarnation, seq) is strictly
-// newer than the last-applied pair and, when it is, advances the high-water
-// mark. incarnation==0 || seq==0 marks a LEGACY (pre-#5706) sender that sends
-// no ordering trailer: admit-always and do NOT advance the mark, mirroring the
-// config-gen gen==0 accept-always compat. The caller holds recvSeqMu.
-func (g *fullSetSeqGuard) admit(incarnation, seq uint64) bool {
-	if incarnation == 0 || seq == 0 {
-		return true // legacy sender — no sequence, accept-always
-	}
-	if g.incarnation == 0 ||
-		incarnation > g.incarnation ||
-		(incarnation == g.incarnation && seq > g.seq) {
-		g.incarnation = incarnation
-		g.seq = seq
-		return true
-	}
-	return false
-}
-
-// reset returns the guard to its zero (accept-next) state. Called from
-// resetRecvGen on a peer bulk re-prime so a rebooted peer with a lower epoch is
-// re-accepted. The caller holds recvSeqMu.
-func (g *fullSetSeqGuard) reset() {
-	g.incarnation = 0
-	g.seq = 0
-}
-
-// stampInstallGenV4 assigns a fresh install generation to a v4 session being
-// sent and records it (keyed by wire key) so the matching delete can echo the
-// exact generation of the install it cancels (#2170 SMR fix #1). It mutates
-// val.Generation in place. A re-send (sweep/bulk) of a live key intentionally
-// bumps the generation: the per-key stored generation only ever climbs, so a
-// journaled delete from before the re-send is strictly older and refused.
-func (s *SessionSync) stampInstallGenV4(key dataplane.SessionKey, val *dataplane.SessionValue) {
-	g := s.nextInstallGen()
-	val.Generation = g
-	// #5274: stamp the admitting config epoch = the config-sync generation
-	// (#3931) this node currently holds. A session still present in the local
-	// table when it is queued has survived this node's own config-apply
-	// clearSessionsForDeletedPolicies sweep, so it is admitted under the
-	// current config; the receiver refuses it only once IT applies a strictly
-	// newer config (lastAppliedConfigGen advances past this epoch). configGen
-	// and lastAppliedConfigGen are the SAME sender→receiver namespace, so the
-	// comparison is meaningful across the HA boundary.
-	val.ConfigEpoch = s.configGenCounter.Load()
-	s.genSentMu.Lock()
-	if s.genSentV4 == nil {
-		s.genSentV4 = make(map[dataplane.SessionKey]uint64)
-	}
-	if !putGenBounded(s.genSentV4, key, g) {
-		s.stats.GenMapOverflow.Add(1)
-	}
-	s.genSentMu.Unlock()
-}
-
-func (s *SessionSync) stampInstallGenV6(key dataplane.SessionKeyV6, val *dataplane.SessionValueV6) {
-	g := s.nextInstallGen()
-	val.Generation = g
-	// #5274: stamp the admitting config epoch (see stampInstallGenV4).
-	val.ConfigEpoch = s.configGenCounter.Load()
-	s.genSentMu.Lock()
-	if s.genSentV6 == nil {
-		s.genSentV6 = make(map[dataplane.SessionKeyV6]uint64)
-	}
-	if !putGenBounded(s.genSentV6, key, g) {
-		s.stats.GenMapOverflow.Add(1)
-	}
-	s.genSentMu.Unlock()
-}
-
-// takeDeleteGenV4 returns the generation a delete for this wire key should
-// carry and evicts the sender-side stamp.
-//
-// #2221: the delete draws a FRESH, strictly-greater generation
-// (nextInstallGen) rather than echoing the install's stamp. The stamp and the
-// sendCh enqueue are not atomic and two producer goroutines (the sweep
-// stamping a live re-send, the delta-drain taking the close) mutate the same
-// key, so a delete can be enqueued onto sendCh BEFORE the install it cancels.
-// On the wire the receiver then applies delete then install with IDENTICAL
-// generations; the receiver guards only refuse a STRICTLY-older operation, so
-// the late install resurrects the closed session (the #2170 residual:
-// stale-RETAIN rather than stale-delete). Drawing a fresh generation that is
-// strictly greater than every prior install of this key makes a delete always
-// out-rank the install it cancels: the receiver's install guard refuses a
-// reordered older install (incoming < the delete tombstone), while a genuinely
-// newer incarnation (re-established + re-stamped by a later sweep) carries an
-// even higher generation and still applies. This composes with #2170: the
-// per-key generation only ever climbs, so a journaled stale delete from before
-// a re-sync is still strictly older than the live entry and refused.
-//
-// A key never installed in this boot (no stamp recorded) returns 0 (legacy
-// fallback → unconditional delete in the apply guard), which is the safe
-// behavior and preserves rolling-upgrade compatibility.
-func (s *SessionSync) takeDeleteGenV4(key dataplane.SessionKey) uint64 {
-	s.genSentMu.Lock()
-	defer s.genSentMu.Unlock()
-	if _, ok := s.genSentV4[key]; !ok {
-		return 0
-	}
-	delete(s.genSentV4, key)
-	return s.nextInstallGen()
-}
-
-func (s *SessionSync) takeDeleteGenV6(key dataplane.SessionKeyV6) uint64 {
-	s.genSentMu.Lock()
-	defer s.genSentMu.Unlock()
-	if _, ok := s.genSentV6[key]; !ok {
-		return 0
-	}
-	delete(s.genSentV6, key)
-	return s.nextInstallGen()
-}
-
-// installGenGuardV4 implements the receiver-side install-side guard (#2170 SMR
-// fix #2): refuse to overwrite a stored entry with a strictly-older-generation
-// install so the per-key stored generation never regresses. Returns the
-// generation to record on a successful apply (the incoming generation, or the
-// preserved stored generation when the incoming one is 0). The bool reports
-// whether the install should proceed.
-func (s *SessionSync) installGenGuardV4(key dataplane.SessionKey, incoming uint64) (record uint64, apply bool) {
-	s.recvGenMu.Lock()
-	defer s.recvGenMu.Unlock()
-	stored, ok := s.recvGenV4[key]
-	if ok && stored != 0 && incoming != 0 && incoming < stored {
-		return 0, false
-	}
-	if incoming == 0 {
-		// Legacy / unknown install: apply but keep the stored generation
-		// (do not roll it back to 0).
-		return stored, true
-	}
-	return incoming, true
-}
-
-func (s *SessionSync) installGenGuardV6(key dataplane.SessionKeyV6, incoming uint64) (record uint64, apply bool) {
-	s.recvGenMu.Lock()
-	defer s.recvGenMu.Unlock()
-	stored, ok := s.recvGenV6[key]
-	if ok && stored != 0 && incoming != 0 && incoming < stored {
-		return 0, false
-	}
-	if incoming == 0 {
-		return stored, true
-	}
-	return incoming, true
-}
-
-// recordInstalledGenV4 stores the per-key generation after a successful
-// install-apply, bounded by genGuardMapCap.
-func (s *SessionSync) recordInstalledGenV4(key dataplane.SessionKey, gen uint64) {
-	if gen == 0 {
-		return
-	}
-	s.recvGenMu.Lock()
-	if s.recvGenV4 == nil {
-		s.recvGenV4 = make(map[dataplane.SessionKey]uint64)
-	}
-	if !putGenBounded(s.recvGenV4, key, gen) {
-		s.stats.GenMapOverflow.Add(1)
-	}
-	s.recvGenMu.Unlock()
-}
-
-func (s *SessionSync) recordInstalledGenV6(key dataplane.SessionKeyV6, gen uint64) {
-	if gen == 0 {
-		return
-	}
-	s.recvGenMu.Lock()
-	if s.recvGenV6 == nil {
-		s.recvGenV6 = make(map[dataplane.SessionKeyV6]uint64)
-	}
-	if !putGenBounded(s.recvGenV6, key, gen) {
-		s.stats.GenMapOverflow.Add(1)
-	}
-	s.recvGenMu.Unlock()
-}
-
-// deleteGenGuardV4 implements the receiver-side delete guard (#2170): a delete
-// is refused only when both the stored and delete generations are non-zero and
-// the delete generation is STRICTLY older than the stored entry's. Equality
-// applies (it is the delete of the very session installed); gen==0 on either
-// side falls back to today's unconditional delete (rolling-upgrade safe). The
-// bool reports whether the delete should proceed.
-//
-// #2221: on an applied delete the stored generation is NOT evicted — it is
-// upgraded to the (strictly-greater, see takeDeleteGenV4) delete generation as
-// a TOMBSTONE. A subsequent reordered install that carries the OLDER generation
-// of the very session this delete cancelled is then refused by installGenGuard*
-// (incoming < the tombstone), so the standby converges to the master's state
-// (session GONE) regardless of install/delete arrival order. A genuinely newer
-// incarnation re-established and re-stamped by a later sweep carries a higher
-// generation and still applies (incoming > tombstone). A gen-0 (legacy) delete
-// evicts (no tombstone to record) — the legacy unconditional path is unchanged.
-// Tombstones are bounded by genGuardMapCap (putGenBounded) and cleared by the
-// bulk barrier (resetRecvGen), so a churning workload cannot grow the map
-// without limit and a cross-boot generation regression is handled at BulkStart.
-func (s *SessionSync) deleteGenGuardV4(key dataplane.SessionKey, deleteGen uint64) bool {
-	s.recvGenMu.Lock()
-	defer s.recvGenMu.Unlock()
-	stored := s.recvGenV4[key]
-	if stored != 0 && deleteGen != 0 && deleteGen < stored {
-		return false
-	}
-	if deleteGen != 0 {
-		// Record the delete generation as a tombstone so a reordered older
-		// install of the cancelled session is refused. Bounded; never clears.
-		if s.recvGenV4 == nil {
-			s.recvGenV4 = make(map[dataplane.SessionKey]uint64)
-		}
-		if !putGenBounded(s.recvGenV4, key, deleteGen) {
-			s.stats.GenMapOverflow.Add(1)
-		}
-	} else {
-		delete(s.recvGenV4, key)
-	}
-	return true
-}
-
-func (s *SessionSync) deleteGenGuardV6(key dataplane.SessionKeyV6, deleteGen uint64) bool {
-	s.recvGenMu.Lock()
-	defer s.recvGenMu.Unlock()
-	stored := s.recvGenV6[key]
-	if stored != 0 && deleteGen != 0 && deleteGen < stored {
-		return false
-	}
-	if deleteGen != 0 {
-		if s.recvGenV6 == nil {
-			s.recvGenV6 = make(map[dataplane.SessionKeyV6]uint64)
-		}
-		if !putGenBounded(s.recvGenV6, key, deleteGen) {
-			s.stats.GenMapOverflow.Add(1)
-		}
-	} else {
-		delete(s.recvGenV6, key)
-	}
-	return true
-}
-
-// resetRecvGen clears the receiver-side stored-generation maps. It is called
-// when the peer begins a fresh bulk transfer (#2198 F2): a reconnecting peer
-// may have REBOOTED, which legitimately restarts its sender genCounter (it is
-// seeded from CLOCK_MONOTONIC nanos, which resets at OS boot). Its bulk
-// re-prime then carries generations that may be LOWER than the generations we
-// stored from the peer's previous boot. Without this reset the install guard
-// would refuse the bulk re-prime as "stale" (stored > incoming) — the inverse
-// of the #2170 bug (stale-RETAIN) — and the cold-start re-prime would silently
-// fail to land.
-//
-// This is safe against opening a stale-delete window: deletes are only acted
-// on after the bulk completes (reconcileStaleSessions runs at BulkEnd), and
-// the bulk re-prime re-establishes the live set (re-recording each key's fresh
-// generation) before any such delete is processed. A delete that arrives
-// mid-bulk for a key the bulk has not yet re-recorded falls back to gen-0
-// (stored==0 after reset) → unconditional, which is the legacy-safe behavior.
-func (s *SessionSync) resetRecvGen() {
-	s.recvGenMu.Lock()
-	s.recvGenV4 = make(map[dataplane.SessionKey]uint64)
-	s.recvGenV6 = make(map[dataplane.SessionKeyV6]uint64)
-	s.recvGenMu.Unlock()
-	// #3931: also reset the last-applied config generation. A reconnecting
-	// peer may have REBOOTED, restarting its monotonic configGenCounter at a
-	// value LOWER than the generation we stored from its previous boot.
-	// Without this reset the config guard (admitConfigGen) would refuse the
-	// reconnect config re-push as stale (last > incoming) — the same
-	// stale-RETAIN inverse the session reset closes (#2198 F2). Resetting to
-	// 0 makes the next config apply unconditionally; it is always the peer's
-	// CURRENT config (pushConfigToPeer sends ShowActive), so the newest
-	// content still wins.
-	s.lastAppliedConfigGen.Store(0)
-	// #5563: reset the received-config high-water alongside the applied mark so
-	// the manual-failover readiness gate's applied<=received invariant holds
-	// across the reset window. The reconnect re-push then re-establishes both
-	// (received on receive, applied on successful apply).
-	s.lastRecvConfigGen.Store(0)
-	// #5706: reset the full-set (IPsec/DHCP) ordering high-water marks for the
-	// same reason. A reconnecting peer that OS-rebooted restarts its monotonic
-	// incarnation LOWER; without this reset the guard would refuse its fresh
-	// full-set re-push (nudged on reconnect) as stale and strand the standby on
-	// the pre-reboot set. Resetting to the zero state admits the next push
-	// unconditionally — it is always the peer's CURRENT set.
-	s.recvSeqMu.Lock()
-	s.ipsecRecvSeq.reset()
-	s.dhcpV4RecvSeq.reset()
-	s.dhcpV6RecvSeq.reset()
-	s.recvSeqMu.Unlock()
-}
-
-// Non-atomicity note (#2198 F3): the apply sequence — guard check
-// (installGenGuard*), PutClusterSynced*, recordInstalledGen* / deleteGenGuard*
-// — does NOT hold recvGenMu across the whole sequence; the mutex is taken
-// independently inside each of installGenGuard*/recordInstalledGen*/
-// deleteGenGuard*. This is safe in production because the receiver apply path
-// for a given peer is single-threaded: messages are decoded and dispatched
-// serially within one receiveLoop goroutine over the single ACTIVE fabric
-// connection (activeConnLocked prefers conn0; conn1 is used only when conn0 is
-// down — never both at once for sends, and the peer sends over one stream). So
-// no two installs/deletes for the SAME key are ever applied concurrently, and
-// the per-key stored generation cannot be interleaved between the guard read
-// and the record write. The standby fabric's receiveLoop exists but the active
-// sender never duplicates a key's traffic across both, so cross-goroutine
-// same-key races do not occur. Holding recvGenMu across the dataplane Put would
-// also serialize unrelated keys and block on dataplane I/O under the lock,
-// which is not worth it for a race that the single-active-fabric invariant
-// already precludes.
-// configEpochStale reports whether a synced session admitted under config
-// epoch `epoch` must be REFUSED because this node has since applied a STRICTLY
-// newer config (#5274). The peer stamps the session with the #3931 config-sync
-// generation it held when it queued the session (stampInstallGen*);
-// lastAppliedConfigGen is the highest config generation THIS node has applied
-// from that same peer, so the two are directly comparable in one
-// sender→receiver namespace. A newer config the peer committed AND this node
-// applied may DENY the session, and this node's clearSessionsForDeletedPolicies
-// sweep for that config already ran — so installing the delayed session would
-// revive a stale PERMIT the config invalidated. epoch==0 (a legacy/pre-#5274
-// peer, or a local-origin entry) disables the check, unconditionally admitting
-// as before (rolling-upgrade safe). The check is authoritative here in the Go
-// cluster layer: the #3931 namespace lives entirely in SessionSync, and the
-// receiver refuses BEFORE forwarding the install to the userspace helper.
-func (s *SessionSync) configEpochStale(epoch uint64) bool {
-	return epoch != 0 && epoch < s.lastAppliedConfigGen.Load()
-}
-
-func (s *SessionSync) installClusterSyncedV4(key dataplane.SessionKey, val dataplane.SessionValue) {
-	if s.sessions == nil {
-		return
-	}
-	record, apply := s.installGenGuardV4(key, val.Generation)
-	if !apply {
-		s.stats.InstallsStaleIgnored.Add(1)
-		slog.Debug("cluster sync: ignored stale-generation v4 install",
-			"incoming_gen", val.Generation)
-		return
-	}
-	if s.configEpochStale(val.ConfigEpoch) {
-		s.stats.SessionsStaleConfigIgnored.Add(1)
-		slog.Debug("cluster sync: ignored stale-config-epoch v4 install (peer moved to a newer config that may deny this session)",
-			"session_config_epoch", val.ConfigEpoch, "applied_config_gen", s.lastAppliedConfigGen.Load())
-		return
-	}
-	if err := s.sessions.PutClusterSyncedV4(key, val); err == nil {
-		s.recordInstalledGenV4(key, record)
-		s.stats.SessionsInstalled.Add(1)
-		s.noteHelperMirrorResult("v4", &s.sessionMirrorWarnedV4, nil)
-		if val.IsReverse == 0 && s.OnForwardSessionInstalled != nil {
-			s.OnForwardSessionInstalled()
-		}
-	} else {
-		s.noteHelperMirrorResult("v4", &s.sessionMirrorWarnedV4, err)
-	}
-}
-
-func (s *SessionSync) installClusterSyncedV6(key dataplane.SessionKeyV6, val dataplane.SessionValueV6) {
-	if s.sessions == nil {
-		return
-	}
-	record, apply := s.installGenGuardV6(key, val.Generation)
-	if !apply {
-		s.stats.InstallsStaleIgnored.Add(1)
-		slog.Debug("cluster sync: ignored stale-generation v6 install",
-			"incoming_gen", val.Generation)
-		return
-	}
-	if s.configEpochStale(val.ConfigEpoch) {
-		s.stats.SessionsStaleConfigIgnored.Add(1)
-		slog.Debug("cluster sync: ignored stale-config-epoch v6 install (peer moved to a newer config that may deny this session)",
-			"session_config_epoch", val.ConfigEpoch, "applied_config_gen", s.lastAppliedConfigGen.Load())
-		return
-	}
-	if err := s.sessions.PutClusterSyncedV6(key, val); err == nil {
-		s.recordInstalledGenV6(key, record)
-		s.stats.SessionsInstalled.Add(1)
-		s.noteHelperMirrorResult("v6", &s.sessionMirrorWarnedV6, nil)
-		if val.IsReverse == 0 && s.OnForwardSessionInstalled != nil {
-			s.OnForwardSessionInstalled()
-		}
-	} else {
-		s.noteHelperMirrorResult("v6", &s.sessionMirrorWarnedV6, err)
-	}
-}
-
-func (s *SessionSync) deleteClusterSyncedV4(key dataplane.SessionKey, deleteGen uint64) {
-	if s.sessions == nil {
-		return
-	}
-	if !s.deleteGenGuardV4(key, deleteGen) {
-		s.stats.DeletesStaleIgnored.Add(1)
-		slog.Debug("cluster sync: ignored stale-generation v4 delete (replacement session survives)",
-			"delete_gen", deleteGen)
-		return
-	}
-	if err := s.sessions.DeleteWithCompanionsV4(key, dataplane.DeleteReasonClusterStale); err != nil {
-		s.stats.Errors.Add(1)
-		slog.Warn("cluster sync: failed to delete v4 session", "err", err)
-	}
-}
-
-func (s *SessionSync) deleteClusterSyncedV6(key dataplane.SessionKeyV6, deleteGen uint64) {
-	if s.sessions == nil {
-		return
-	}
-	if !s.deleteGenGuardV6(key, deleteGen) {
-		s.stats.DeletesStaleIgnored.Add(1)
-		slog.Debug("cluster sync: ignored stale-generation v6 delete (replacement session survives)",
-			"delete_gen", deleteGen)
-		return
-	}
-	if err := s.sessions.DeleteWithCompanionsV6(key, dataplane.DeleteReasonClusterStale); err != nil {
-		s.stats.Errors.Add(1)
-		slog.Warn("cluster sync: failed to delete v6 session", "err", err)
-	}
-}
-
 func shouldInitiateFabricDial(localAddr, peerAddr string) bool {
 	local, err := netip.ParseAddrPort(localAddr)
 	if err != nil {
